feat: add 'manual mode'
When "manual mode" is enabled, Shepherd will manually pull locally the latest
tagged image from the registry, check if its newer compared to the one in use
by the service (by comparing their unique identifiers, called _digests_), and
in such case it will try and force Docker to update the service with it.

This feature has been initially developed to have a temporary workaround/alternative
until the "pulled images with none tag" issue is fixed:
 SEE: https://github.com/moby/moby/issues/28908

It could also be used to isolate and troubleshoot issues that one thinks may be
related to the "update process", by setting some services in "manual mode" and
monitoring their status, compared to the ones updated fully using the built-in
logic and not "forcing".

Includes:
- Add "manual mode" feature to the main script.
- Update the README.
- Add more comments to the code.