Created a logging function and added a verbose mode
Log function which takes the log and whether it is considered a verbose mode log. If it is a verbose log and verbose mode is false it will NOT print the log. If verbose mode is true it will print all logs. Also using a timezone environment variable to set the timezone file at /etc/timezone.